feat: Update button labels for replies and comments
The code changes in Replies.vue and Comments.vue update the button labels for submitting replies and comments. The "Submit" button has been changed to "Reply" in Replies.vue and "Comment" in Comments.vue. This improves the clarity and user-friendliness of the interface by using more descriptive labels for the actions.